Jasper

+++ src/main/java/com/revature/service/TechnologyService.java
@@ -1,7 +1,8 @@
 package com.revature.service;
 
+import java.util.List;
 import java.util.Optional;
-import java.util.Set;
+
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -9,7 +10,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
 import com.revature.model.Technology;
 import com.revature.repository.TechnologyDao;
 import com.revature.exceptions.TechnologyNotFoundException;
@@ -21,13 +21,13 @@ public class TechnologyService {
 	private Logger log = LoggerFactory.getLogger(this.getClass());
 
 	@Transactional(readOnly = true)
-	public Optional<Set<Technology>> findAll() {
-		return tDao.getAll();
+	public List<Technology> findAll() {
+		return tDao.findAll();
 	}
 
 	@Transactional(readOnly = true)
 	public Optional<Technology> getById(int id) {
-		return tDao.getById(id);
+		return tDao.findById(id);
 	}
 	
 	@Transactional(readOnly = true)
@@ -45,22 +45,28 @@ public class TechnologyService {
 	@Transactional(propagation = Propagation.REQUIRED)
 	public void deleteTechnology(int id) {
 		try {
+			if (id < 0) {
+				throw new IllegalArgumentException("TechId cannot be less than 0");
+			}
 			tDao.deleteById(id);
+			System.out.println("catch hit");
 		} catch (IllegalArgumentException e) {
 			log.warn("ID can't be pulled.");
 		}
 	}
 	
 	@Transactional(propagation = Propagation.REQUIRED)
-	public void update(Technology tech)
+	public Technology update(Technology tech)
 	{
 		try
 		{
-			tDao.save(tech);
+		return tDao.save(tech);
 		}
 		catch(IllegalArgumentException e)
+
 		{
 			log.warn("Technology can't be found, update failed.");
+			return null;
 		}
 	}
 }
